+++ src/factory.ts
@@ -1,37 +1,46 @@
+import { Util, DomEvent, DomUtil, Browser, PolyUtil } from 'leaflet';
+
 import Element, { ElementProps } from '@components/Factory/Element';
+
 import ElementPortal, {
   ElementPortalProps,
   ElementPortalRef,
 } from '@components/Factory/ElementPortal';
+
 import ElementRoot, {
   ElementRootProps,
   ElementRootRef,
 } from '@components/Factory/ElementRoot';
-import { Util, DomEvent, DomUtil, Browser, PolyUtil } from 'leaflet';
+
 import {
   CustomEvent,
   EventHandlers,
   UseElementEventsHook,
   useElementEvents,
 } from '@hooks/useElementEvents';
+
 import {
   Constructor,
   UseElementFactoryHook,
   Validation,
   useElementFactory,
 } from '@hooks/useElementFactory';
+
 import {
   AddCallback,
   UseElementLifeCycleHook,
   useElementLifeCycle,
 } from '@hooks/useElementLifeCycle';
+
 import { useElementParent } from '@hooks/useElementParent';
+
 import {
   UseElementUpdateHook,
   useElementUpdate,
 } from '@hooks/useElementUpdate';
-import { BaseFactoryMethods } from '@utils/interfaces';
+
 import Validator from '@helpers/validator';
+import { BaseFactoryMethods } from '@utils/interfaces';
 
 export {
   Element,