+++ src/index.ts
@@ -1,25 +1,31 @@
 import Map from './components/Map';
+
 import Icon from './components/Icon';
 import Popup from './components/Popup';
 import Marker from './components/Marker';
 import DivIcon from './components/DivIcon';
 import Tooltip from './components/Tooltip';
-import TileLayer from './components/TileLayer';
-import ClusterGroup from './components/ClusterGroup';
-import CanvasGroup from './components/CanvasGroup';
 import CircleMarker from './components/CircleMarker';
-import ControlContainer, { FullScreen, Zoom } from './components/Control';
+
+import LayerGroup from './components/LayerGroup';
+import CanvasGroup from './components/CanvasGroup';
+import ClusterGroup from './components/ClusterGroup';
+import FeatureGroup from './components/FeatureGroup';
+
 import GeoJSON from './components/GeoJSON';
 import Polygon from './components/Polygon';
 import Polyline from './components/Polyline';
-import FeatureGroup from './components/FeatureGroup';
-import LayerGroup from './components/LayerGroup';
 import Rectangle from './components/Rectangle';
+
 import SVGOverlay from './components/SVGOverlay';
 import ImageOverlay from './components/ImageOverlay';
 import VideoOverlay from './components/VideoOverlay';
+
+import TileLayer from './components/TileLayer';
 import TileLayerWMS from './components/TileLayerWMS';
+
 import AttributionControl from './components/Control/Attribution';
+import ControlContainer, { FullScreen, Zoom } from './components/Control';
 
 import 'leaflet/dist/leaflet.css';
 
@@ -50,6 +56,14 @@ export {
   Zoom,
 };
 
+// hooks
+
+export * from './hooks/useMap';
+export * from './hooks/useZoom';
+export * from './hooks/useGeoSearch';
+export * from './hooks/useFullscreen';
+export * from './hooks/useRoutingMachine';
+
 export type * from './components/Map';
 export type * from './components/Icon';
 export type * from './components/Popup';
@@ -75,8 +89,5 @@ export type * from './components/Control/Attribution';
 export type * from './components/Control/Container';
 export type * from './components/Control/Fullscreen';
 export type * from './components/Control/Zoom';
-export * from './hooks/useFullscreen';
-export * from './hooks/useMap';
-export * from './hooks/useZoom';
 
 export type { SVG } from './components/SVGOverlay';