+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+[1.0.2] - 
+
+### Fixed Popup error: "cannot read properties of undefined: reading _source"
+### Improvements to Factory
+### fixed useElementUpdate to update props state even when the next value is nullish
+### Added Leaflet Routing Machine Hook
+### Added Leaflet Geosearch Hook
+### Added /helpers path
+### Small general improvements
+
 [1.0.1] - 2024-06-19
 
 ### CSS File import included